fix: fix lambda trigger crud product quickbook online

+++ lambdas/trigger_crud_product_quickbook_online/main.py
@@ -5,23 +5,19 @@ from datetime import datetime
 import urllib3
 import json
 
-# def decode_data(data_encode: str) -> str:
-#     string_data = base64.b64decode(data_encode).decode("utf-8")
-#     return string_data
-
 def lambda_handler(event: Dict[str, Any], context):
     host = os.environ['API_HOST']
-    # rent("Response:", response.data)
     for record in event['Records']:
         print(record)
         method = None
         record_data = json.loads(record["body"])
-        if record_data.get("action").upper() == "CREATE":
-            method = "POST"
-        elif record_data.get("action").upper() == "UPDATE":
-            method = "PATCH"
-        elif record_data.get("action").upper() == "DELETE":
-            method = "DELETE"
+        if record_data.get("action") is not None:
+            if record_data.get("action").upper() == "CREATE":
+                method = "POST"
+            elif record_data.get("action").upper() == "UPDATE":
+                method = "PATCH"
+            elif record_data.get("action").upper() == "DELETE":
+                method = "DELETE"
 
         encoded_data = json.dumps(record_data).encode("utf-8")
